1. Update to fix Keil Compiler 6 issue 2. Fixed some warning in sample codes

+++ SampleCode/StdDriver/TIMER_FreeCountingMode/main.c
@@ -35,7 +35,7 @@ void TMR0_IRQHandler(void)
         else
         {
             /* TIMER0 clock source = PCLK0 = HCLK / 2 = HIRC / 2 */
-            printf("Input frequency is %dHz\n", (__HIRC/2) / (t1 - t0));
+            printf("Input frequency is %dHz\n", (int32_t)(__HIRC/2) / (t1 - t0));
         }
     }
     else
